refactor(#743): keep files under the 1500-line maintainability guardrail
The RC1 additions pushed execution_flow.py to 1531 and two quality_gates test
parts over 1500.

- Extract the blocked-resume setup-phase decision and the resume-setup-failure
  re-block into _blocked_resume_setup_phase_names / _reblock_on_resume_setup_failure
  in state_ops.py, wired via the executor mixin — execution_flow.py stays a thin
  shell (1499 lines).
- Relocate the #743 quality-gate tests (unowned_protected_paths, workspace.yml
  no-op message) into a dedicated tests/unit/control/test_quality_gates_743.py so
  the existing part files drop back under the limit.

No behavior change; the extracted helpers are covered by the existing RC1 tests.
