+++ tests/unit/control/test_quality_gates_parts/test_quality_gates_part_004.py
@@ -732,41 +732,6 @@ fail_under = 80
 
 
 @pytest.mark.unit
-def test_violation_message_flags_workspace_yml_tier_and_final_gate_noop() -> None:
-    """A .awf/workspace.yml edit must warn the tier/final_gate edit is a no-op (#743).
-
-    ``requested_tier`` is set at dispatch and ``final_gate`` is fixed at provision
-    time, so editing them in the protected file has no effect. The message must
-    say so, so agents stop editing the file to try to escalate AWF validation.
-    """
-    violations = find_protected_quality_gate_changes(
-        changed_paths=[".awf/workspace.yml"],
-        owned_paths=[],
-    )
-
-    message = quality_gate_violation_message(violations)
-
-    assert ".awf/workspace.yml" in message
-    assert "requested_tier" in message
-    assert "final_gate" in message
-    assert "dispatch" in message
-
-
-@pytest.mark.unit
-def test_violation_message_omits_workspace_yml_note_for_other_files() -> None:
-    """The tier/final_gate note only appears for .awf/workspace.yml edits."""
-    violations = find_protected_quality_gate_changes(
-        changed_paths=["pytest.ini"],
-        owned_paths=[],
-    )
-
-    message = quality_gate_violation_message(violations)
-
-    assert "pytest.ini" in message
-    assert "requested_tier" not in message
-
-
-@pytest.mark.unit
 def test_missing_protected_file_diff_blocks_conservatively() -> None:
     violations = find_protected_quality_gate_changes(
         changed_paths=["pyproject.toml"],
